requirements.txt
numpy==2.2.2
pandas==2.2.3
snowflake-connector-python==3.12.4
python-dotenv==1.0.1
pytest==8.3.4
requests==2.32.3
snowflake-connector-python[pandas]